+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## TBD
+* Flag error if an asset change tries to subtract more funds from an asset than exist
+
 ## 0.3.0
 * Render only scenarios with year >= currentYear, and provide an `--all` flag to render all scenarios
 * Fix bug that warns user about stale gov constants when the latest gov constants are from a future year